+++ packages/api-client/src/v2/features/search/hooks/use-search-index.ts
@@ -1,17 +1,16 @@
+import type { Item } from '@scalar/sidebar'
 import type { OpenApiDocument } from '@scalar/workspace-store/schemas/v3.1/strict/openapi-document'
-import type { FuseResult } from 'fuse.js'
 import { type MaybeRefOrGetter, computed, ref, toValue } from 'vue'
 
 import { createFuseInstance } from '../helpers/create-fuse-instance'
 import { createSearchIndex } from '../helpers/create-search-index'
-import type { FuseData } from '../types'
 
 const MAX_SEARCH_RESULTS = 25
 
 /**
- * Creates the search index from an OpenAPI document.
+ * Creates the search index from multiple OpenAPI documents
  */
-export function useSearchIndex(documents: MaybeRefOrGetter<OpenApiDocument[]>) {
+export const useSearchIndex = (documents: MaybeRefOrGetter<OpenApiDocument[]>) => {
   /** When the document changes we replace the search index */
   const fuse = computed(() => {
     const instance = createFuseInstance()
@@ -21,24 +20,23 @@ export function useSearchIndex(documents: MaybeRefOrGetter<OpenApiDocument[]>) {
 
   const query = ref<string>('')
 
-  const results = computed(() => {
+  const results = computed<Item[] | null>(() => {
     if (query.value.length !== 0) {
-      return fuse.value.search(query.value, {
+      const fuseResults = fuse.value.search(query.value, {
         limit: MAX_SEARCH_RESULTS,
+      })
+
+      // Lets only show operations for now to match previous behavior
+      return fuseResults.flatMap((result) => {
+        if (result.item.entry.type !== 'operation') {
+          return []
+        }
+
+        return result.item.entry
       })
     }
 
-    // @ts-expect-error - _docs is a private property
-    const allEntries: FuseData[] = fuse.value._docs
-
-    // Show a few entries as a placeholder
-    return allEntries.slice(0, MAX_SEARCH_RESULTS).map(
-      (item: FuseData, index: number) =>
-        ({
-          item,
-          refIndex: index,
-        }) satisfies FuseResult<FuseData>,
-    )
+    return null
   })
 
   return {