+++ packages/api-reference/src/components/AgentScalar/AgentScalarChatInterface.vue
@@ -2,13 +2,16 @@
 import { Chat } from '@scalar/agent-chat'
 import { type ApiReferenceConfigurationWithSource } from '@scalar/types/api-reference'
 import type { WorkspaceStore } from '@scalar/workspace-store/client'
+import type { Ref } from 'vue'
 
 import { API_BASE_URL, DASHBOARD_URL, REGISTRY_URL } from '@/consts/urls'
 
-const { agentConfig, workspaceStore } = defineProps<{
-  agentConfig: ApiReferenceConfigurationWithSource['agent']
-  workspaceStore: WorkspaceStore
-}>()
+const { agentScalarConfiguration, workspaceStore, prefilledMessage } =
+  defineProps<{
+    agentScalarConfiguration: ApiReferenceConfigurationWithSource['agent']
+    workspaceStore: WorkspaceStore
+    prefilledMessage?: Ref<string>
+  }>()
 </script>
 
 <template>
@@ -16,8 +19,13 @@ const { agentConfig, workspaceStore } = defineProps<{
     :baseUrl="API_BASE_URL"
     :dashboardUrl="DASHBOARD_URL"
     :getActiveDocumentJson="() => workspaceStore.exportActiveDocument('json')!"
-    :getAgentKey="agentConfig?.key ? () => agentConfig?.key : undefined"
-    :mode="agentConfig?.key ? 'full' : 'preview'"
+    :getAgentKey="
+      agentScalarConfiguration?.key
+        ? () => agentScalarConfiguration?.key ?? ''
+        : undefined
+    "
+    :mode="agentScalarConfiguration?.key ? 'full' : 'preview'"
+    :prefilledMessage="prefilledMessage"
     :registryDocuments="[]"
     :registryUrl="REGISTRY_URL" />
 </template>