+++ packages/openapi-parser/src/utils/is-object.ts
@@ -1,4 +1,0 @@
-/**
- * Check if the given value is an object
- */
-export const isObject = (obj: any) => typeof obj === 'object' && !Array.isArray(obj) && obj !== null